git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@13013 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickCore/image.c
@@ -1673,7 +1673,7 @@ MagickExport MagickBooleanType IsHighDynamicRangeImage(const Image *image,
           continue;
         pixel=(double) p[i];
         if ((pixel < 0.0) || (pixel > QuantumRange) ||
-            (pixel != (QuantumAny) pixel))
+            (pixel != (double) ((QuantumAny) pixel)))
           break;
       }
       p+=GetPixelChannels(image);
